Adds functionality to fab in PlaylistTabFragment
