+++ app/src/main/java/com/musicplayer/SocyMusic/data/base/PlaylistDao.java
@@ -1,6 +1,7 @@
 package com.musicplayer.SocyMusic.data.base;
 
 import androidx.room.Dao;
+import androidx.room.Delete;
 import androidx.room.Insert;
 import androidx.room.Query;
 import androidx.room.Update;
@@ -28,4 +29,6 @@ public interface PlaylistDao {
     @Update
     void update(Playlist playlist);
 
+    @Delete
+    void delete(Playlist playlist);
 }
