+++ app/src/main/java/com/musicplayer/SocyMusic/ui/playlist/PlaylistActivity.java
@@ -77,7 +77,7 @@ public class PlaylistActivity extends PlayerFragmentHost implements PlaylistFrag
     }
 
     @Override
-    public void onNewPlaylist() {
+    public void onNewPlaylist(Playlist newPlaylist) {
 
     }
